pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "familydoc-ai"
version = "0.1.0"
description = "A simple RAG-based AI therapist for the ReadyTensor course."
authors = [
    { name = "Egor Zubenko", email = "ezubenko@gmail.com" },
]
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "typer[rich]>=0.12.3",
    "langchain>=0.2.1",
    "langchain-openai>=0.1.7",
    "langchain-huggingface>=0.0.3",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.47.0",
    "langchain_community",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.5",
    "pytest>=8.2.1",
    "pytest-asyncio>=0.23.5",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = ["src"]
